chore: split sidebar components into separate files
