+++ packages/app/src/app/pages/Dashboard/Sidebar/BottomMessages/UserUpgradeToTeamPro.tsx
@@ -3,6 +3,6 @@ import { Text } from '@codesandbox/components';
 
 export const UserUpgradeToTeamPro = () => (
   <Text css={{ color: '#999', fontWeight: 400, fontSize: 12 }}>
-    Contact team admin to upgrade for the full Codesandbox Experience.{' '}
+    Contact team admin to upgrade for the full Codesandbox Experience.
   </Text>
 );
